server/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
transformers>=4.40.0
torch>=2.0.0
PyJWT>=2.9.0
bcrypt>=4.0.0
python-multipart>=0.0.9
safetensors>=0.4.0
